refactor: seguindo boas práticas de nomenclatura para index dentro dos components
