+++ src/frontend/src/App.tsx
@@ -1,9 +1,9 @@
 import { useState } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Toaster } from "sonner";
-import Login from "./pages/Login/Login";
-import Home from "./pages/Home/Home";
-import NotFound from "./pages/NotFound/NotFound";
+import Login from "./pages/Login";
+import Home from "./pages/Home";
+import NotFound from "./pages/NotFound";
 
 const App = () => {
   const [loggedIn, setLoggedIn] = useState(false);
